+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## xel.js 1.4.0  not yet released
 
 * Implement MROUND(n, m)
+* Implement MROUND2(n, m)
 
 
 ## xel.js 1.3.0  released 2024-04-05